Added TiledMapping for declarative property mapping and cleaned up TiledObject
- Made code conform to new C++11 standards
- Replaced "foreach" with "for" loop
- Swapped "0" with "nullptr"
- Switched TiledObject from QQuickItem to QObject
- Added TiledBackground to get background properties
- Separated private API from public API

+++ src/viewport.cpp
@@ -215,14 +215,14 @@ void Viewport::setScene(Scene *scene)
 {
     m_scene = scene;
     scene->setParentItem(this);
-    setContentWidth(scene->width());
-    setContentHeight(scene->height());
+    setContentWidth(static_cast<float>(scene->width()));
+    setContentHeight(static_cast<float>(scene->height()));
     setVisible(true);
     updateMaxOffsets();
 }
 
 void Viewport::updateMaxOffsets()
 {
-    m_maxXOffset = m_contentWidth - width();
-    m_maxYOffset = m_contentHeight - height();
+    m_maxXOffset = m_contentWidth - static_cast<float>(width());
+    m_maxYOffset = m_contentHeight - static_cast<float>(height());
 }
